Improve the way the increased skill is found
Thanks to Jakob for the suggestions and the pattern I copied from him!
This is almost certainly faster since it uses `tes3mp` methods which
are written in C++.

I also added some placeholder debug messages for leveling, to help me
work out if that stuff is functional or not.