Fix decay happening way too fast
Due to following the mwscript a bit too closely, I ended up forcing a
very early decay for pretty much any skill above the hardcoded 15
limit.

I've fixed that by just calculating the `decayMemory` variable to what
it needs to be (and I checked that value with the mwscript version to
be sure).  Along with a few other related and not related changes,
this now enables decay to work as expected - the first skill drop
won't happen for several hours of gameplay.