Clean up event conditionals, fix leveling bug
Introduced along with the level cap mechanic, the conditional added in
that commit prevented leveling up from happening unless there was a
level cap at all.

This is now done a bit better, with a `canLevel()` method that returns
true or false if the player can level (based on configs and other
things) that is tested against.

I also changed how the various event methods tested for execution.  In
an effort to reduce the loop depth, they now return immediately versus
doing an else.  It seems cleaner.