Fix not factoring training sessions, ignore level progress
It turns out that level progress is how the vanilla game knows to do
it's level up thing.  MWscript NCGD stops that from happening by
making level progress always 0, so let's do that too.  All level
progress value handling is now done in `OnPlayerSkill` so we can keep
it 0.

I've also added a block to ensure that skills are properly calculated
after a trainer is used; this falls under the same logic block as when
they die or go to jail.

And lastly: I removed several comments and made some tiny
optimizations with variable setting.