Fix `modHealth` behavior to prevent stat reset (#2)
The basic idea is to switch the flow of data. Before, this function used
the data saved to server as the source of truth. Now, it uses the client
data as the source of truth. Before, it updated the saved server data,
and loaded data from there into the client. Now, it updates client data
and server data directly, then sends stats to all clients.

 1. We now read stat values from the client rather than the server.
 2. If health difference is less than one, prevent packet spam.
 3. Refactored the `getHealthGetRatio` logic to run only once, then
    removed the function entirely.
 4. Removed fortify health check, it seems to work fine without it..?
 5. Renamed variables with `new` and `old` prefix for clarity.
 6. Inlined `temp` variable manipulations.